requirements.txt
Django==4.1.6
djangorestframework==3.14.0
environs==9.5.0
drf-yasg==1.21.4
